requirements.txt
aniso8601==9.0.1
attrs==23.2.0
Automat==22.10.0
Babel==2.14.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
furl==2.1.3
greenlet==3.0.3
hyperlink==21.0.0
idna==3.6
importlib_resources==6.4.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
orderedmultidict==1.0.1
packaging==24.0
passlib==1.7.4
python-dateutil==2.9.0.post0
pytz==2024.1
redo==2.0.4
requests==2.31.0
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.29
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
zope.interface==6.2
